+++ src/app/services/book.service.ts
@@ -1,12 +1,34 @@
 import { inject, Injectable } from '@angular/core';
-import { collection, Firestore, getDocs, addDoc, updateDoc, deleteDoc, doc, DocumentReference, DocumentData } from '@angular/fire/firestore';
+import {
+  collection,
+  Firestore,
+  getDocs,
+  addDoc,
+  updateDoc,
+  deleteDoc,
+  doc,
+  DocumentReference,
+  DocumentData,
+} from '@angular/fire/firestore';
 import { BookDetailsModel } from '../shared/book-details/book/book.types';
-import { BehaviorSubject, filter, from, map, Observable, of, switchMap } from 'rxjs';
+import {
+  BehaviorSubject,
+  filter,
+  from,
+  map,
+  Observable,
+  of,
+  switchMap,
+  throwError,
+} from 'rxjs';
+import { AuthService } from '../auth/auth.service';
 
 @Injectable({ providedIn: 'root' })
 export class BookService {
   firestore = inject(Firestore);
   private readonly COLLECTION_NAME = 'books';
+
+  constructor(private authService: AuthService) {}
 
   private booksSubject = new BehaviorSubject<BookDetailsModel[] | null>(null);
   books$ = this.booksSubject.asObservable();
@@ -39,25 +61,47 @@ export class BookService {
     );
   }
 
-  createBook(book: Omit<BookDetailsModel, 'id'>): Observable<DocumentReference<DocumentData>> {
-    const itemCollection = collection(this.firestore, this.COLLECTION_NAME);
-    return from(addDoc(itemCollection, { ...book })).pipe(
-      switchMap((docRef) => this.fetchBooksObservable().pipe(map(() => docRef)))
-    );
+  createBook(
+    book: Omit<BookDetailsModel, 'id'>
+  ): Observable<DocumentReference<DocumentData>> {
+    if (this.authService.isAuthenticated()) {
+      const itemCollection = collection(this.firestore, this.COLLECTION_NAME);
+      return from(addDoc(itemCollection, { ...book })).pipe(
+        switchMap((docRef) =>
+          this.fetchBooksObservable().pipe(map(() => docRef))
+        )
+      );
+    } else {
+      return throwError(() => new Error('User not authenticated'));
+    }
   }
 
   updateBook(bookId: string, book: BookDetailsModel): Observable<void> {
-    const bookDocRef = doc(this.firestore, `${this.COLLECTION_NAME}/${bookId}`);
-    return from(updateDoc(bookDocRef, { ...book })).pipe(
-      switchMap(() => this.fetchBooksObservable())
-    );
+    if (this.authService.isAuthenticated()) {
+      const bookDocRef = doc(
+        this.firestore,
+        `${this.COLLECTION_NAME}/${bookId}`
+      );
+      return from(updateDoc(bookDocRef, { ...book })).pipe(
+        switchMap(() => this.fetchBooksObservable())
+      );
+    } else {
+      return throwError(() => new Error('User not authenticated'));
+    }
   }
 
   deleteBook(bookId: string): Observable<void> {
-    const bookDocRef = doc(this.firestore, `${this.COLLECTION_NAME}/${bookId}`);
-    return from(deleteDoc(bookDocRef)).pipe(
-      switchMap(() => this.fetchBooksObservable())
-    );
+    if (this.authService.isAuthenticated()) {
+      const bookDocRef = doc(
+        this.firestore,
+        `${this.COLLECTION_NAME}/${bookId}`
+      );
+      return from(deleteDoc(bookDocRef)).pipe(
+        switchMap(() => this.fetchBooksObservable())
+      );
+    } else {
+      return throwError(() => new Error('User not authenticated'));
+    }
   }
 
   getMilyenSzinLennelBook(): Observable<BookDetailsModel> {
